# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.CreateFortuneInfoRequest:
    properties:
      birth_day:
        example: 1
        maximum: 31
        minimum: 1
        type: integer
      birth_hour:
        example: 12
        maximum: 23
        minimum: 0
        type: integer
      birth_minute:
        example: 0
        maximum: 59
        minimum: 0
        type: integer
      birth_month:
        example: 1
        maximum: 12
        minimum: 1
        type: integer
      birth_place:
        example: 서울
        type: string
      birth_year:
        example: 2000
        maximum: 2100
        minimum: 1900
        type: integer
      unknown_time:
        example: false
        type: boolean
    required:
    - birth_day
    - birth_month
    - birth_place
    - birth_year
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  handler.RecordsResponse:
    properties:
      records:
        items: {}
        type: array
    type: object
...
  handler.SimilarUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/handler.UserWithScore'
        type: array
    type: object
  handler.TodayFortuneResponse:
    properties:
      fortune:
        example: 오늘은 새로운 시작에 좋은 날입니다. 자신감을 가지고 도전해보세요.
        type: string
    type: object
  handler.UserWithScore:
    properties:
      similarity_score:
        example: 85.5
        type: number
      user: {}
    type: object
...
  models.FortuneInfo:
    properties:
      birth_day:
        example: 1
        type: integer
      birth_hour:
        example: 12
        type: integer
      birth_minute:
        example: 0
        type: integer
      birth_month:
        example: 1
        type: integer
      birth_place:
        example: 서울
        type: string
      birth_year:
        example: 2000
        type: integer
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      day_earthly_branch:
        example: 子
        type: string
      day_heavenly_stem:
        example: 甲
        type: string
      hour_earthly_branch:
        example: 子
        type: string
      hour_heavenly_stem:
        example: 甲
        type: string
      id:
        example: 1
        type: integer
      month_earthly_branch:
        example: 寅
        type: string
      month_heavenly_stem:
        example: 戊
        type: string
      unknown_time:
        example: false
        type: boolean
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 1
        type: integer
      year_earthly_branch:
        example: 子
        type: string
      year_heavenly_stem:
        example: 庚
        type: string
    type: object
host: localhost:8080
info:
  title: DoTheFortune API
  version: "1.0"
paths:
...
  /fortune/info:
    get:
      consumes:
      - application/json
      description: 현재 사용자의 등록된 사주 정보를 조회합니다. 사주 정보가 없으면 404 오류가 반환됩니다.
      produces:
      - application/json
      responses:
        "200":
          description: 사주 정보 조회 성공
          schema:
            $ref: '#/definitions/models.FortuneInfo'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 사주 정보를 찾을 수 없음
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 사주 정보 조회
      tags:
      - fortune
    post:
      consumes:
      - application/json
      description: 사용자의 사주 정보를 등록하거나 수정합니다. 생년월일, 출생 시각, 출생지를 입력받아 사주를 계산하고 저장합니다.
        출생 시각을 모를 경우 unknown_time을 true로 설정하면 자동으로 12시 0분으로 설정됩니다.
      parameters:
      - description: 사주 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateFortuneInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 사주 정보 저장 성공
          schema:
            $ref: '#/definitions/models.FortuneInfo'
        "400":
          description: 잘못된 요청 (필수 필드 누락, 잘못된 날짜 형식 등)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 사주 정보 등록/수정
      tags:
      - fortune
  /fortune/similar:
    get:
      consumes:
      - application/json
      description: 현재 사용자와 유사한 사주를 가진 다른 사용자들을 찾아 반환합니다. 일간 천간 또는 지지를 기준으로 유사도를 계산하며,
        유사도 점수(0-100)와 함께 반환됩니다.
      parameters:
      - default: 10
        description: 반환할 최대 사용자 수
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 유사 사용자 목록
          schema:
            $ref: '#/definitions/handler.SimilarUsersResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 유사 사주 친구 찾기
      tags:
      - fortune
  /fortune/today:
    get:
      consumes:
      - application/json
      description: 사용자의 사주 정보를 기반으로 오늘의 운세를 한 줄로 제공합니다. 사주 정보가 등록되어 있어야 하며, 조회 시 기록이
        자동으로 저장됩니다.
      produces:
      - application/json
      responses:
        "200":
          description: 오늘의 운세 조회 성공
          schema:
            $ref: '#/definitions/handler.TodayFortuneResponse'
        "400":
          description: 사주 정보가 등록되지 않음
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 오늘의 운세 조회
      tags:
      - fortune
  /records/:
    get:
      consumes:
      - application/json
      description: 사용자의 최근 운세 기록을 조회합니다. 오늘의 운세, 궁합 결과, AI 이미지 등 모든 타입의 기록이 최신순으로
        반환됩니다.
      parameters:
      - default: 20
        description: 반환할 최대 기록 수
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 기록 목록 조회 성공
          schema:
            $ref: '#/definitions/handler.RecordsResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 최근 기록 조회
      tags:
      - records
  /records/{type}:
    get:
      consumes:
      - application/json
      description: 특정 타입의 운세 기록만 필터링하여 조회합니다. 타입은 'today_fortune', 'compatibility',
        'ai_spouse' 등이 있습니다.
      parameters:
      - description: 기록 타입 (today_fortune, compatibility, ai_spouse 등)
        in: path
        name: type
        required: true
        type: string
      - default: 20
        description: 반환할 최대 기록 수
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 타입별 기록 목록
          schema:
            $ref: '#/definitions/handler.RecordsResponse'
        "401":
          description: 인증 실패
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 서버 내부 오류
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 타입별 기록 조회
      tags:
      - records
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 사용한 인증. "Bearer " 뒤에 공백을 두고 JWT 토큰을 입력하세요. 예: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"